% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add.buffer}
\alias{add.buffer}
\title{Add Buffer to Time Series}
\usage{
add.buffer(TS, n)
}
\arguments{
\item{TS}{Time series data.}

\item{n}{Buffer size.}
}
\value{
Time series with added buffer.
}
\description{
Adds a buffer to a time series using auto.arima.
}
