% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{do.synth}
\alias{do.synth}
\title{Apply the Synth Method to Time-Series Data (Internal)}
\usage{
do.synth(
  df,
  dep.var,
  dependent.id,
  predictors,
  special.predictors,
  time.predictors.prior,
  time.optimize.ssr
)
}
\arguments{
\item{df}{A data frame containing the data for analysis.}

\item{dep.var}{A string indicating the dependent variable column name.}

\item{dependent.id}{An identifier for the treated unit or case.}

\item{predictors}{A list or vector of predictor variable names.}

\item{special.predictors}{A list or vector of special predictor names.}

\item{time.predictors.prior}{Time periods for which to use the predictors.}

\item{time.optimize.ssr}{Time period to optimize the SSR.}
}
\value{
A list containing the actual values (\code{value}), average values (\code{average}),
and the synthetic control predictions (\code{synthetic}).
}
\description{
This function performs the synthetic control method on provided data. It prepares the data
and fits a synthetic control model using the \code{synth} package. This is an internal helper
function and is not intended to be directly used by package users.
}
\keyword{internal}
