% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{dsc}
\alias{dsc}
\title{Dynamic Synthetic Control (DSC) method}
\usage{
dsc(
  data,
  start.time,
  end.time,
  treat.time,
  dependent,
  k = 4,
  filter.width = 5,
  buffer = 0,
  norm.method = "t",
  match.method = "fixed",
  step.pattern1 = dtw::symmetricP1,
  step.pattern2 = dtw::asymmetricP2,
  plot.figures = FALSE,
  n.burn = 3,
  ma = 3,
  ma.na = "original",
  dist.quant = 1,
  n.IQR = 3,
  window.type = "none",
  default.margin = 3,
  n.q = 1,
  n.r = 1,
  parallel = TRUE,
  rescale = TRUE,
  dependent.id,
  predictors,
  special.predictors,
  time.predictors.prior,
  time.optimize.ssr
)
}
\arguments{
\item{data}{A data frame containing the observational data.}

\item{start.time}{Starting time for the analysis.}

\item{end.time}{Ending time for the analysis.}

\item{treat.time}{Treatment time.}

\item{dependent}{The dependent variable name in the dataset.}

\item{k}{Integer, number of control units used for dynamic time warping. Default is 4.}

\item{filter.width}{Integer, width of the filter. Default is 5.}

\item{buffer}{Integer, buffer for time series alignment. Default is 0.}

\item{norm.method}{Method for normalization. Default is "t".}

\item{match.method}{Method for matching. Default is "fixed".}

\item{step.pattern1}{Step pattern for the DTW. Default is dtw::symmetricP1.}

\item{step.pattern2}{Alternative step pattern for DTW. Default is dtw::asymmetricP2.}

\item{plot.figures}{Logical, if TRUE plots will be generated. Default is FALSE.}

\item{n.burn}{Integer, number of initial time periods to disregard. Default is 3.}

\item{ma}{Integer, moving average length. Default is 3.}

\item{ma.na}{Method to handle missing values in moving average. Default is "original".}

\item{dist.quant}{Numeric, quantile for distance measure. Default is 1.}

\item{n.IQR}{Numeric, factor for IQR in outlier detection. Default is 3.}

\item{window.type}{Type of window for DTW. Default is "none".}

\item{default.margin}{Default margin size. Default is 3.}

\item{n.q}{Integer, number of synthetic controls to use. Default is 1.}

\item{n.r}{Integer, number of predictors to use. Default is 1.}

\item{parallel}{Logical, if TRUE parallel processing will be enabled. Default is TRUE.}

\item{rescale}{Logical, if TRUE data will be rescaled. Default is TRUE.}

\item{dependent.id}{Numeric, ID of the dependent unit.}

\item{predictors}{List, names of predictor variables.}

\item{special.predictors}{List, names of special predictor variables.}

\item{time.predictors.prior}{List, names of time predictor variables for the prior period.}

\item{time.optimize.ssr}{List, names of time predictor variables for the SSR optimization.}
}
\value{
A list containing results of the synthetic control analysis.
}
\description{
The DSC method combines ideas from the synthetic control method and dynamic time warping.
This function processes the given data, applies time-series preprocessing, and then
computes the synthetic control using dynamic time warping.
}
\examples{
\dontrun{
library(dsc)

# Load the Basque dataset from the Synth package
data(basque, package = "Synth")
data <- basque

# Rename relevant columns for clarity
colnames(data)[1:4] <- c("id", "unit", "time", "value")

# Compute additional variables
data$invest_ratio <- data$invest / data$value
data$value_raw <- data$value

# Define special predictors for the model
special_preds <- expression(list(
  list(dep.var, 1960:1969, c("mean")),
  list("invest_ratio", 1964:1969, c("mean")),
  list("popdens", 1969, c("mean")),
  list("sec.agriculture", 1961:1969, c("mean")),
  list("sec.energy", 1961:1969, c("mean")),
  list("sec.industry", 1961:1969, c("mean")),
  list("sec.construction", 1961:1969, c("mean")),
  list("sec.services.venta", 1961:1969, c("mean")),
  list("sec.services.nonventa", 1961:1969, c("mean")),
  list("school.illit", 1964:1969, c("mean")),
  list("school.prim", 1964:1969, c("mean")),
  list("school.med", 1964:1969, c("mean")),
  list("school.high", 1964:1969, c("mean")),
  list("school.post.high", 1964:1969, c("mean"))
))

# Execute the DSC analysis
result <- dsc(
  data = data,
  start.time = 1955,
  end.time = 1997,
  treat.time = 1970,
  dependent = "Basque Country (Pais Vasco)",
  predictors = NULL,
  parallel = TRUE,
  special.predictors = special_preds,
  time.predictors.prior = 1955:1969,
  time.optimize.ssr = 1955:1969
)
}

}
