% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_bc.R
\name{DSC_bc}
\alias{DSC_bc}
\title{Barycenters in the DSC Method}
\usage{
DSC_bc(controls, weights, evgrid = seq(from = 0, to = 1, length.out = 1001))
}
\arguments{
\item{controls}{a list or matrix:
in list form, each list element contains a vector of observations for the given control unit;
in matrix form, each column corresponds to one unit and each row is one observation.}

\item{weights}{a vector of weights for each control unit; can be calculated by \code{\link[DSC]{DSC_weights}}.}

\item{evgrid}{a vector of gridpoints used to evaluate quantile functions.}
}
\value{
\code{DSC_bc} returns a list containing the following components:
\item{\code{controls.q}}{a list of evaluated control unit quantile functions.}
\item{\code{bc}}{a vector of evaluated barycenter.}
}
\description{
Function for computing the barycenter in the DSC method at a specific time period.
}
\examples{
#simulated data from Gaussian mixture
#ex_gmm() calls the simulated data
#detail can be found by ??ex_gmm
#obtaining the optimal weights
DSC_res=DSC_weights(ex_gmm()$control,ex_gmm()$target, M=100)
weights=DSC_res[[1]]
#obtaining the barycenter
DSC_bc(ex_gmm()$control,weights,seq(from=0,to=1,length.out=1001))
}
