% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_plot.R
\name{DSC_plot}
\alias{DSC_plot}
\title{Create a Plot for the Target Distribution and Barycenter}
\usage{
DSC_plot(
  evgrid = seq(from = 0, to = 1, length.out = 1001),
  bc,
  target,
  y_name = "y",
  x_name = "x"
)
}
\arguments{
\item{evgrid}{a vector of gridpoints used to evaluate quantile functions.}

\item{bc}{a vector of evaluated barycenter; can be calculated by \code{\link[DSC]{DSC_bc}}.}

\item{target}{a vector of observations.}

\item{y_name}{a string for the title of the y-axis.}

\item{x_name}{a string for the title of the x-axis.}
}
\description{
Create a plot for the target distribution (in black) and barycenter (in red).
}
\examples{
#simulated data from Gaussian mixture
#ex_gmm() calls the simulated data
#detail can be found by ??ex_gmm
#obtaining the optimal weights
DSC_res=DSC_weights(ex_gmm()$control,ex_gmm()$target, M=100)
weights=DSC_res[[1]]
DSC_res2=DSC_bc(ex_gmm()$control,weights,seq(from=0,to=1,length.out=1001))
DSC_plot(bc=DSC_res2[[2]],target=ex_gmm()$target)
}
