% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_weights.R
\name{DSC_weights}
\alias{DSC_weights}
\title{Weights in the DSC Method}
\usage{
DSC_weights(controls, target, M = 500, choose_solver = "SCS")
}
\arguments{
\item{controls}{a list or matrix:
in list form, each list element contains a vector of observations for the given control unit;
in matrix form, each column corresponds to one unit and each row is one observation.}

\item{target}{a vector of observations.}

\item{M}{an integer specifying the number of draws from the uniform distribution for approximating the integral.}

\item{choose_solver}{a solver for the optimization problem; see \code{\link[CVXR]{installed_solvers}} in CVXR for more options.}
}
\value{
\code{DSC_weights} returns a list containing the following components:
\item{\code{weights}}{a matrix of the optimal weights.}
\item{\code{distance}}{a float of the squared Wasserstein distance between the target and the corresponding barycenter.}
}
\description{
Function for obtaining the weights in the DSC method at a specific time period.
}
\examples{
#simulated data from Gaussian mixture
#ex_gmm() calls the simulated data
#detail can be found by ??ex_gmm
DSC_weights(ex_gmm()$control,ex_gmm()$target, M=100)
}
