% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_gmm.R
\name{ex_gmm}
\alias{ex_gmm}
\title{Simulated Data from Gaussian Mixture}
\usage{
ex_gmm()
}
\value{
\item{\code{target}}{a vector.}
\item{\code{control}}{a matrix.}
}
\description{
Example data for \code{DSC_weights}, \code{DSC_bc}, \code{DSC_plot}, \code{DSC_CI}.
There are 30 control units, each has a distribution of a mixture of 3 Gaussian distributions.
The target has a distribution of a mixture of 4 Gaussian distributions.
}
