% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_normal.R
\name{ex_normal}
\alias{ex_normal}
\title{Simulated data from independent normal distributions}
\usage{
ex_normal()
}
\value{
\item{\code{target}}{a list.}
\item{\code{control}}{a list.}
}
\description{
Example data for \code{DSC_main}, \code{DSC_per}.
There are 10 periods. Post-treatment periods are 1:5.
There are 30 control units.
In each period, each control unit has a distribution of normal distribution with randomly generated mean in unif(0,1),
ramdomly gnerated variance in unif(0.8,1.2).
The target has a distribution of N(0,1) in t=1:5, and N(2,1) in t=6:10.
}
