% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFDTW.R
\name{first.dtw}
\alias{first.dtw}
\title{Dynamic Time Warping: First Phase}
\usage{
first.dtw(
  x,
  y,
  t.treat,
  buffer = 10,
  norm.method = "t",
  match.method = "fixed",
  step.pattern = dtw::symmetricP2,
  window.type = "none",
  window.size = NULL,
  plot.figures = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{t.treat}{Treatment time.}

\item{buffer}{Buffer size. Default is 10.}

\item{norm.method}{Normalization method. Default is "t".}

\item{match.method}{Matching method. Default is "fixed".}

\item{step.pattern}{Step pattern to use. Default is \code{dtw::symmetricP2}.}

\item{window.type}{Type of window. Default is "none".}

\item{window.size}{Optional size for the window. Default is NULL.}

\item{plot.figures}{Logical, whether to plot figures. Default is FALSE.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A list containing various outputs from the DTW process.
}
\description{
This function calculates the dynamic time warping for the first phase.
}
\keyword{internal}
