% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{minmax.normalize}
\alias{minmax.normalize}
\title{Min-Max Normalization}
\usage{
minmax.normalize(data, reference = NULL)
}
\arguments{
\item{data}{A numeric vector to be normalized.}

\item{reference}{An optional reference for normalization.}
}
\value{
Normalized data.
}
\description{
Normalizes the data using min-max scaling.
}
