% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normalize}
\alias{normalize}
\title{General Normalization Function}
\usage{
normalize(data, norm.method, reference = NULL)
}
\arguments{
\item{data}{A numeric vector to be normalized.}

\item{norm.method}{A string that specifies the normalization method.}

\item{reference}{An optional reference for normalization.}
}
\value{
Normalized data.
}
\description{
Applies the specified normalization to the data.
}
