% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot.synth}
\alias{plot.synth}
\title{Plot Synthetic Control}
\usage{
\method{plot}{synth}(
  res.synth.raw,
  res.synth.TFDTW,
  dependent,
  start.time,
  end.time,
  treat.time,
  file.name,
  legend.pos = c(0.3, 0.3)
)
}
\arguments{
\item{res.synth.raw}{Results for the raw synthetic control.}

\item{res.synth.TFDTW}{Results for the TFDTW synthetic control.}

\item{dependent}{Dependent variable.}

\item{start.time}{Start time.}

\item{end.time}{End time.}

\item{treat.time}{Treatment time.}

\item{y.limits}{Y axis limits.}
}
\value{
A ggplot object.
}
\description{
Visualizes the synthetic control results.
}
