% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot.warped}
\alias{plot.warped}
\title{Plot Warped Data}
\usage{
\method{plot}{warped}(
  unit,
  dependent,
  t.treat,
  y.raw,
  x.raw,
  x.warped,
  legend.pos = c(0.3, 0.3)
)
}
\arguments{
\item{unit}{Data unit.}

\item{dependent}{Dependent variable.}

\item{t.treat}{Treatment time.}

\item{y.raw}{Raw y values.}

\item{x.raw}{Raw x values.}

\item{x.warped}{Warped x values.}

\item{legend.pos}{Position of the legend.}
}
\value{
A ggplot object.
}
\description{
Visualizes the warped data.
}
