% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{preprocessing}
\alias{preprocessing}
\title{Pre-process Data}
\usage{
preprocessing(
  data,
  filter.width = 5,
  norm.method = "t",
  n.poly = 3,
  n.deri = 2,
  plot.data = FALSE
)
}
\arguments{
\item{data}{The data to be processed.}

\item{filter.width}{Width of the filter.}

\item{norm.method}{Normalization method.}

\item{n.poly}{Degree of the polynomial.}

\item{n.deri}{Order of the derivative.}

\item{plot.data}{Logical, indicating whether to plot the data.}
}
\value{
Processed data.
}
\description{
Applies several pre-processing steps to the data.
}
