% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{RefTooShort}
\alias{RefTooShort}
\title{Check if Reference is Short in DTW}
\usage{
RefTooShort(
  query,
  reference,
  step.pattern = dtw::symmetricP2,
  window.type = "none",
  window.size = NULL
)
}
\arguments{
\item{query}{The query series.}

\item{reference}{The reference series.}

\item{step.pattern}{Step pattern for DTW.}

\item{window.type}{Type of window for DTW.}

\item{window.size}{Size of the window for DTW.}
}
\value{
Logical indicating if reference is too short.
}
\description{
Checks if the reference series is too short for dynamic time warping.
}
