% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{RemoveOutliers}
\alias{RemoveOutliers}
\title{Remove Outliers in Weight Matrix}
\usage{
RemoveOutliers(data, n.IQR = 3)
}
\arguments{
\item{data}{Numeric data.}

\item{n.IQR}{Multiplier for the interquartile range.}
}
\value{
Data with outliers removed.
}
\description{
Removes outliers from the data based on interquartile range.
}
