% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stack.warped}
\alias{stack.warped}
\title{Stack Warped Data Figures}
\usage{
\method{stack}{warped}(fig.list, file.name, ncol = 3)
}
\arguments{
\item{fig.list}{List of ggplot objects.}

\item{file.name}{Name of the output file.}

\item{ncol}{Number of columns for the stacked figures.}
}
\description{
Stacks multiple warped data figures.
}
