% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{t.normalize}
\alias{t.normalize}
\title{Standard Z-score Normalization}
\usage{
\method{t}{normalize}(data, reference = NULL)
}
\arguments{
\item{data}{A numeric vector to be normalized.}

\item{reference}{An optional reference for normalization.}
}
\value{
Normalized data.
}
\description{
Normalizes the data using z-score.
}
