% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFDTW.R
\name{TFDTW}
\alias{TFDTW}
\title{Two-Step Dynamic Time Warping}
\usage{
TFDTW(
  x,
  y,
  k,
  t.treat,
  buffer,
  norm.method = "t",
  match.method = "fixed",
  step.pattern1 = dtw::symmetricP2,
  step.pattern2 = dtw::asymmetricP2,
  plot.figures = FALSE,
  n.burn = 3,
  ma = 3,
  ma.na = "original",
  dist.quant = 1,
  n.IQR = 3,
  window.type = "none",
  default.margin = 3,
  n.q = 1,
  n.r = 1,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{k}{An integer.}

\item{t.treat}{Treatment time.}

\item{buffer}{Buffer size.}

\item{norm.method}{Normalization method. Default is "t".}

\item{match.method}{Matching method. Default is "fixed".}

\item{step.pattern1}{Step pattern for the first phase. Default is \code{dtw::symmetricP2}.}

\item{step.pattern2}{Step pattern for the second phase. Default is \code{dtw::asymmetricP2}.}

\item{plot.figures}{Logical, whether to plot figures. Default is FALSE.}

\item{n.burn}{Burn-in size. Default is 3.}

\item{ma}{Moving average window size. Default is 3.}

\item{ma.na}{Method for handling NA in moving average. Default is "original".}

\item{dist.quant}{Distance quantile. Default is 1.}

\item{n.IQR}{Number of IQR. Default is 3.}

\item{window.type}{Type of window. Default is "none".}

\item{default.margin}{Default margin size. Default is 3.}

\item{n.q}{An integer, default 1.}

\item{n.r}{An integer, default 1.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A list containing various outputs from the two-step DTW process.
}
\description{
This function performs a two-step dynamic time warping.
}
\keyword{internal}
