% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warp2weight}
\alias{warp2weight}
\title{Convert Warping Path to Weight}
\usage{
warp2weight(W)
}
\arguments{
\item{W}{The warping path.}
}
\value{
Weights.
}
\description{
Transforms a warping path to weights.
}
