% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warpWITHpath}
\alias{warpWITHpath}
\title{Warp Time Series with Warping Path}
\usage{
warpWITHpath(w, ts)
}
\arguments{
\item{w}{Warping path.}

\item{ts}{Time series to be warped.}
}
\value{
Warped time series.
}
\description{
Warps a time series using a specified warping path.
}
