% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warpWITHweight}
\alias{warpWITHweight}
\title{Warp Time Series with Weights}
\usage{
warpWITHweight(ts, weight)
}
\arguments{
\item{ts}{Time series to be warped.}

\item{weight}{Weights for warping.}
}
\value{
Warped time series.
}
\description{
Warps a time series using the provided weights.
}
